Inline functions need to be defined at declaration

+++ 6502/main.cpp
@@ -5,5 +5,5 @@ int main() {
 	EMU6502::CpuData cpu;
 	uint8_t val = 235;
 	std::cout << "should be 235: ";
-	std::cout << cpu.AddrImm(val) << std::endl;
+	std::cout << (int)cpu.AddrImm(val) << std::endl;
 }
